# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/admin/products": {
      "get": {
        "tags": [
          "Admin Panel"
        ],
        "summary": "Fetches data for products page",
        "description": "You need to execute /auth/login as admin first",
        "produces": [
          "json"
        ],
        "responses": {
          "401": {
            "description": "Unauthorized: No token provided or invalid token"
          },
          "403": {
            "description": "Unauthorized: Only admins have access"
          },
          "500": {
            "description": "Internal Server Error: Error fetching data for products page"
          }
        }
      }
    },
...
    "/admin/roles": {
      "get": {
        "tags": [
          "Admin Panel"
        ],
        "summary": "Fetches data for roles page",
        "description": "You need to execute /auth/login as admin first",
        "produces": [
          "json"
        ],
        "responses": {
          "401": {
            "description": "Unauthorized: No token provided or invalid token"
          },
          "403": {
            "description": "Unauthorized: Only admins have access"
          },
          "500": {
            "description": "Internal Server Error: Error fetching data for roles page"
          }
        }
      }
    },
...
    "/products/": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "View all the products in the database",
        "description": "",
        "produces": [
          "json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not found: No products found"
          },
          "500": {
            "description": "Internal Server Error: Error when getting products"
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Adds a new product.",
        "description": "You need to execute /auth/login as admin first",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/PostProduct"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created: Product created successfully"
          },
          "400": {
            "description": "Bad request: Invalid or missing properties"
          },
          "401": {
            "description": "Unauthorized: No token provided or invalid token"
          },
          "403": {
            "description": "Unauthorized: Only admins have access"
          },
          "500": {
            "description": "Internal Server Error: An error occurred while trying to add product"
          }
        }
      }
    },
    "/products/{id}": {
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Update a product by ID",
        "description": "You need to execute /auth/login as admin first",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UpdateProduct"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created: Product updated successfully"
          },
          "400": {
            "description": "Bad Request: Invalid properties"
          },
          "401": {
            "description": "Unauthorized: No token provided or invalid token"
          },
          "403": {
            "description": "Unauthorized: Only admins have access"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error: An error occurred while trying to update product"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "Delete a product by ID",
        "description": "NOTE: This is a soft delete. <br><br/> <br> You need to execute /auth/login as admin first <br/>",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK: Product deleted successfully"
          },
          "400": {
            "description": "Bad Request: Product is already deleted"
          },
          "401": {
            "description": "Unauthorized: No token provided or invalid token"
          },
          "403": {
            "description": "Unauthorized: Only admins have access"
          },
          "404": {
            "description": "Product not found"
          },
          "500": {
            "description": "Internal Server Error: An error occurred while trying to update product"
          }
        }
      }
    },
...
    "/roles/": {
      "get": {
        "tags": [
          "Users and Roles"
        ],
        "summary": "Get all Role alternatives",
        "description": "",
        "produces": [
          "json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error: An error occurred while trying to get roles"
          }
        }
      }
    }
  },
  "definitions": {
...
    "PostProduct": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "iPhone 2030 Model"
        },
        "imgurl": {
          "type": "string",
          "example": "example-url.png"
        },
        "description": {
          "type": "string",
          "example": "Do all on your phone"
        },
        "price": {
          "type": "number",
          "example": 11999
        },
        "quantity": {
          "type": "number",
          "example": 40
        },
        "brandId": {
          "type": "number",
          "example": 1
        },
        "categoryId": {
          "type": "number",
          "example": 1
        }
      },
      "required": [
        "name",
        "imgurl",
        "description",
        "price",
        "quantity",
        "brandId",
        "categoryId"
      ]
    },
    "UpdateProduct": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "newiPhoneName"
        },
        "description": {
          "type": "string",
          "example": "This iPhone have a new name and price"
        },
        "price": {
          "type": "number",
          "example": 9999
        }
      }
    },
...
  }
}